+++ R/estimate_slopes.R
@@ -16,71 +16,23 @@
 #' various examples, tutorials and use cases.
 #'
 #' @param trend A character indicating the name of the variable for which to
-#' compute the slopes.
+#' compute the slopes. To get marginal effects at specific values, use
+#' `trend="<variable>"` along with the `by` argument, e.g.
+#' `by="<variable>=c(1, 3, 5)"`, or a combination of `by` and `length`, for
+#' instance, `by="<variable>", length=30`. To calculate average marginal
+#' effects over a range of values, use `trend="<variable>=seq(1, 3, 0.1)"` (or
+#' similar) and omit the variable provided in `trend` from the `by` argument.
 #' @param p_adjust The p-values adjustment method for frequentist multiple
 #' comparisons. For `estimate_slopes()`, multiple comparison only occurs for
 #' Johnson-Neyman intervals, i.e. in case of interactions with two numeric
-#' predictors (one specified in `trend`, one in `by`). In this case, the `"esarey"`
-#' option is recommended, but `p_adjust` can also be one of `"none"` (default),
-#' `"hochberg"`, `"hommel"`, `"bonferroni"`, `"BH"`, `"BY"`, `"fdr"`, `"tukey"`,
-#' `"sidak"`, or `"holm"`.
+#' predictors (one specified in `trend`, one in `by`). In this case, the
+#' `"esarey"` option is recommended, but `p_adjust` can also be one of `"none"`
+#' (default), `"hochberg"`, `"hommel"`, `"bonferroni"`, `"BH"`, `"BY"`, `"fdr"`,
+#' `"tukey"`, `"sidak"`, or `"holm"`.
 #' @inheritParams estimate_means
 #' @inheritParams parameters::model_parameters.default
 #'
-#' @details
-#' The [estimate_slopes()], [estimate_means()] and [estimate_contrasts()]
-#' functions are forming a group, as they are all based on *marginal*
-#' estimations (estimations based on a model). All three are built on the
-#' **emmeans** or **marginaleffects** package (depending on the `backend`
-#' argument), so reading its documentation (for instance [emmeans::emmeans()],
-#' [emmeans::emtrends()] or this [website](https://marginaleffects.com/)) is
-#' recommended to understand the idea behind these types of procedures.
-#'
-#' - Model-based **predictions** is the basis for all that follows. Indeed,
-#' the first thing to understand is how models can be used to make predictions
-#' (see [estimate_link()]). This corresponds to the predicted response (or
-#' "outcome variable") given specific predictor values of the predictors (i.e.,
-#' given a specific data configuration). This is why the concept of [`reference
-#' grid()`][insight::get_datagrid()] is so important for direct predictions.
-#'
-#' - **Marginal "means"**, obtained via [estimate_means()], are an extension
-#' of such predictions, allowing to "average" (collapse) some of the predictors,
-#' to obtain the average response value at a specific predictors configuration.
-#' This is typically used when some of the predictors of interest are factors.
-#' Indeed, the parameters of the model will usually give you the intercept value
-#' and then the "effect" of each factor level (how different it is from the
-#' intercept). Marginal means can be used to directly give you the mean value of
-#' the response variable at all the levels of a factor. Moreover, it can also be
-#' used to control, or average over predictors, which is useful in the case of
-#' multiple predictors with or without interactions.
-#'
-#' - **Marginal contrasts**, obtained via [estimate_contrasts()], are
-#' themselves at extension of marginal means, in that they allow to investigate
-#' the difference (i.e., the contrast) between the marginal means. This is,
-#' again, often used to get all pairwise differences between all levels of a
-#' factor. It works also for continuous predictors, for instance one could also
-#' be interested in whether the difference at two extremes of a continuous
-#' predictor is significant.
-#'
-#' - Finally, **marginal effects**, obtained via [estimate_slopes()], are
-#' different in that their focus is not values on the response variable, but the
-#' model's parameters. The idea is to assess the effect of a predictor at a
-#' specific configuration of the other predictors. This is relevant in the case
-#' of interactions or non-linear relationships, when the effect of a predictor
-#' variable changes depending on the other predictors. Moreover, these effects
-#' can also be "averaged" over other predictors, to get for instance the
-#' "general trend" of a predictor over different factor levels.
-#'
-#' **Example:** Let's imagine the following model `lm(y ~ condition * x)` where
-#' `condition` is a factor with 3 levels A, B and C and `x` a continuous
-#' variable (like age for example). One idea is to see how this model performs,
-#' and compare the actual response y to the one predicted by the model (using
-#' [estimate_expectation()]). Another idea is evaluate the average mean at each of
-#' the condition's levels (using [estimate_means()]), which can be useful to
-#' visualize them. Another possibility is to evaluate the difference between
-#' these levels (using [estimate_contrasts()]). Finally, one could also estimate
-#' the effect of x averaged over all conditions, or instead within each
-#' condition (`using [estimate_slopes]`).
+#' @inherit estimate_means details
 #'
 #' @inheritSection estimate_means Predictions and contrasts at meaningful values (data grids)
 #'
@@ -118,11 +70,28 @@
 #' )
 #' summary(slopes)
 #' plot(slopes)
+#'
+#' # marginal effects, grouped by Species, at different values of Petal.Length
+#' estimate_slopes(model,
+#'   trend = "Petal.Length",
+#'   by = c("Petal.Length", "Species"), length = 10
+#' )
+#'
+#' # marginal effects at different values of Petal.Length
+#' estimate_slopes(model, trend = "Petal.Length", by = "Petal.Length", length = 10)
+#'
+#' # marginal effects at very specific values of Petal.Length
+#' estimate_slopes(model, trend = "Petal.Length", by = "Petal.Length=c(1, 3, 5)")
+#'
+#' # average marginal effects of Petal.Length,
+#' # just for the trend within a certain range
+#' estimate_slopes(model, trend = "Petal.Length=seq(2, 4, 0.01)")
 #' }
 #' @export
 estimate_slopes <- function(model,
                             trend = NULL,
                             by = NULL,
+                            predict = NULL,
                             ci = 0.95,
                             p_adjust = "none",
                             transform = NULL,
@@ -140,6 +109,7 @@ estimate_slopes <- function(model,
     estimated <- get_emtrends(
       model,
       trend = trend,
+      predict = predict,
       by = by,
       keep_iterations = keep_iterations,
       verbose = verbose,
@@ -152,6 +122,7 @@ estimate_slopes <- function(model,
       model,
       trend = trend,
       by = by,
+      predict = predict,
       ci = ci,
       p_adjust = p_adjust,
       transform = transform,